requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
boto3==1.34.144
botocore==1.34.144
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Deprecated==1.2.14
dnspython==2.6.1
easyocr==1.7.2
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.37.0
img2table==1.4.1
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jose==1.0.0
lazy_loader==0.4
llvmlite==0.44.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msal==1.31.0
multidict==6.1.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==2.0.1
Office365-REST-Python-Client==2.5.13
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
packaging==25.0
pandas==2.2.2
passlib==1.7.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==10.4.0
polars==1.29.0
prettytable==3.15.1
propcache==0.2.1
protobuf==4.25.3
psycopg2-binary==2.9.9
pyarrow==20.0.0
pyasn1==0.6.0
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.25.4
pyngrok==7.2.3
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
rsa==4.9
s3cmd==2.4.0
s3transfer==0.10.2
scikit-image==0.25.2
scipy==1.15.2
sendgrid==6.11.0
sentry-sdk==2.11.0
setuptools==71.0.1
shapely==2.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.21
starkbank-ecdsa==2.2.0
starlette==0.37.2
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.3.30
torch==2.7.0
torchvision==0.22.0
twilio==9.4.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
XlsxWriter==3.2.3
yarl==1.18.3
zipp==3.19.2
markdown>=3.5.2
